Fix slow SSR calls
Our SSR calls were pretty consistently taking ~0.3s to render a simple
page. We add a benchmarking harness at the rust level to test the core
library performance - and this returned in 8.4ms. The problem was
upstream at the TimeoutProcess level. Wrapping any logic in a
separate process and yielding it back to the main process seemed
to create about a 0.3 overhead. We switch the approach to a thread
to fix this rendering time.