Switch to annotated response typehint for passthrough/sideeffect
AsyncIterator and Iterator couldn't be properly passed and typehinted
into the response_model argument of @passthrough. Decorating as an
explicit response annotation provides more flexibility and has the
benefit of allowing static typehinting to verify the actual function
conforms to the advertised spec.

As part of this change, we still support the old response_model
parameter but we flag a warning that it will be deprecated in a future
release.