Remove unnecessary controller-name in response schema
Since the JS layer owns the state update logic, the server manipulating
the action payloads to include the controller name was redundent and
negatively affected the client API contract by forcing a
response.Controller.passthrough to get the raw payload.

We revert this change in preference to having each endpoint return its
raw payload and letting the action caller distinguish which global state
payload to update.