Support polymorphic superclass sideeffects
Sideeffect response models are unique because they require a concrete
RenderBase of a mounted controller, but the functions themselves can
still be defined at the level of the superclass. Previously we only had
a 1:1 relationship between an action function's metadata and the
response model - but in cases of sideeffects this resulted in a last-win
assignment of the render model to the `sideeffect` property. Our global
controllers typehinted incorrectly in this situation.

This commit solves the issue by registering action render models for
each concrete controller that's registered. Global controllers will do a
union of each identity, whereas the controller owned implementation
itself will only use its more specific sideeffect render.