Fix nondeterministic crash on pku flags
When V8 is initialized on a non-main thread on CPU architectures that
have PKU enabled, rendering can occasionally crash during memory
allocation:

```
Thread 36 "python" received signal SIGSEGV, Segmentation fault.
[Switching to Thread 0x7fff69e006c0 (LWP 134254)]
0x00007ffff2f728b1 in v8::internal::FreeListManyCachedFastPathBase::Allocate(unsigned long, unsigned long*, v8::internal::AllocationOrigin) ()
```

When this happens a SIGSEGV is thrown in the V8 isolate itself, which is
caught by neither Rust nor Python so the application exits with no
further context about the issue.